Change timestamp to double as workaround for bug in explode component in Grasshopper

+++ Diffing_oM/Delta.cs
@@ -48,7 +48,7 @@ namespace BH.oM.Diffing
 
         public Stream DiffingStream { get; } = null;
 
-        public long Timestamp { get; }
+        public double Timestamp { get; }
 
         public string Author { get; }
